fix: do not require a deployment ID to evolve with CLI (#618)
This change updates the deployment workflow in the CLI to not require a
previous deployment ID when evolving. The previous deployment ID is
worked out on the server-side so it should be safe to call evolve
without needing to look up deployments first.

The previous behavior was making day 0 harder with `gram push --method
merge` because that calls EvolveDeployment under the hood and it was
failing for newly created Gram projects that no deployments.