fix(dashboard): remote OAS fetching via proxying the request (#985)
Proxy the request for remote resources through a gram endpoint to play
nicely with CORs and CSP settings.

The behavior here changes a little bit, too, as it'd be weird to proxy
the request back to the client and have the client send it right back to
the server, so we're proxying the remote urls to the server and turning
them into assets then-and-there.